--- name-of-your-project/src/main.jsx
@@ -1,22 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import {
-  createBrowserRouter,
-  RouterProvider,
-  Route,
-} from "react-router-dom";
-import "./index.css";
-import Root from "./routes/root";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root/>,
-  },
-]);
-
-ReactDOM.createRoot(document.getElementById("root")).render(
-  <React.StrictMode>
-    <RouterProvider router={router} />
-  </React.StrictMode>
-);
