+++ package.json
@@ -17,9 +17,12 @@
     "axios": "^1.1.3",
     "classnames": "^2.3.2",
     "cypress": "^10.10.0",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.2",
+    "sort-by": "^1.2.0",
     "vite-plugin-svgr": "^2.2.1"
   },
   "devDependencies": {
