+++ src/routes/api-delegation.tsx
@@ -1,0 +1,10 @@
+import { Page, PageContent, PageHeader } from "@altinn/altinn-design-system";
+
+export default function ApiDelegation() {
+    return(
+        <Page>
+            <PageHeader>PageHeader</PageHeader>
+            <PageContent>PageContent</PageContent>
+        </Page>
+    )
+};
