--- src/routes/expenses.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Expenses() {
-    return (
-      <main style={{ padding: "1rem 0" }}>
-        <h2>Expenses</h2>
-      </main>
-    );
-  }
