--- src/routes/invoices.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Invoices() {
-    return (
-      <main style={{ padding: "1rem 0" }}>
-        <h2>Invoices</h2>
-      </main>
-    );
-}
