+++ src/routes/root.jsx
@@ -29,10 +29,7 @@ export default function Root() {
           <nav>
             <ul>
               <li>
-                <a href={`contacts/1`}>Your Name</a>
-              </li>
-              <li>
-                <a href={`contacts/2`}>Your Friend</a>
+                <a href={`api-delegation/1`}>Page 1</a>
               </li>
             </ul>
           </nav>
